feat(resolve): creator plugin with dynamic widget
