+++ pype/plugins/resolve/create/create_shot_clip.py
@@ -1,13 +1,6 @@
 from pprint import pformat
 from pype.hosts import resolve
 from pype.hosts.resolve import lib
-import re
-
-
-def camel_case_split(text):
-    matches = re.finditer(
-        '.+?(?:(?<=[a-z])(?=[A-Z])|(?<=[A-Z])(?=[A-Z][a-z])|$)', text)
-    return " ".join([str(m.group(0)).capitalize() for m in matches])
 
 
 class CreateShotClip(resolve.Creator):
@@ -18,34 +11,8 @@ class CreateShotClip(resolve.Creator):
     icon = "film"
     defaults = ["Main"]
 
+    gui_name = "Define sequencial rename"
     presets = None
-
-    # widget
-    layout = [{
-        "type": "QLabel",
-        "label": "Define sequencial rename"
-    }]
-
-    def add_presets_to_layout(self, data):
-        for k, v in data.items():
-            if isinstance(v, dict):
-                self.layout.append({
-                    "type": "QLabel",
-                    "label": camel_case_split(k)
-                })
-                self.add_presets_to_layout(v)
-            elif isinstance(v, str):
-                self.layout.append({
-                    "type": "QLineEdit",
-                    "label": camel_case_split(k),
-                    "setText": v
-                })
-            elif isinstance(v, int):
-                self.layout.append({
-                    "type": "QSpinBox",
-                    "label": camel_case_split(k),
-                    "setValue": v
-                })
 
     def process(self):
         print(f"__ selected_clips: {self.selected}")
@@ -53,9 +20,7 @@ class CreateShotClip(resolve.Creator):
         if len(self.selected) < 1:
             return
 
-        self.add_presets_to_layout(self.presets)
-
-        widget = self.widget(self.layout)
+        widget = self.widget(self.gui_name, self.presets)
         widget.exec_()
 
         print(widget.result)
